pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "srcc-faq-bot"
version = "0.1.0"
description = "囲碁ロボット(SRCC) コールセンター用 FAQ・用語集ボット"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "llama-index-core>=0.10.50",
    "llama-index-vector-stores-pinecone>=0.1.9",
    "llama-index-embeddings-openai>=0.1.11",
    "llama-index-llms-anthropic>=0.1.15",
    "llama-index-postprocessor-cohere-rerank>=0.1.7",
    "rank-bm25>=0.2.2",
    "upstash-redis>=1.1.0",
    "anthropic>=0.28",
    "openai>=1.35",
    "pinecone-client>=3.2",
    "python-dotenv>=1.0",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "mangum>=0.17",        # AWS Lambda ASGI adapter
    "httpx>=0.27",
    "playwright>=1.40",
    "beautifulsoup4>=4.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "httpx>=0.27",
    "ruff>=0.4",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"
